Remove config layer and move some config class in cli/config And some import fixes related type.py & type classes
