+++ alma_sbom/formats/cyclonedx/models.py
@@ -9,7 +9,7 @@ from cyclonedx.output.xml import XmlV1Dot4
 
 from alma_sbom import constants
 from alma_sbom.data.models import Package, Build
-from alma_sbom.config.config import SbomFileFormatType
+from alma_sbom.type import SbomFileFormatType
 from ..document import Document as AlmasbomDocument
 from .component import component_from_package, component_from_build
